requirements.txt
APScheduler==2.1.2
autopep8==1.5.4
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dialogflow==1.1.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
google-api-core==1.22.2
google-api-python-client==1.11.0
google-auth==1.21.1
google-auth-httplib2==0.0.4
google-oauth2-tool==0.0.3
googleapis-common-protos==1.52.0
grpcio==1.31.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
oauth2client==4.1.3
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
toml==0.10.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
xmltodict==0.12.0
